bug: View와 ViewController 분리 (#61)
