+++ DietLog-MVVM/View/Exercise/View/Cells/ExerciseTableViewCell.swift
