+++ DietLog-MVVM/View/Exercise/View/ExerciseEditView.swift
@@ -1,49 +1,44 @@
 //
-//  ExerciseEditViewController.swift
+//  ExerciseEditView.swift
 //  DietLog-MVVM
 //
-//  Created by Jooyeon Kang on 2024/04/24.
+//  Created by Jooyeon Kang on 2024/05/04.
 //
 
 import UIKit
-import RxSwift
-import RxCocoa
 
-enum ExerciseEditOption: Int {
-    case URL, category, memo
+protocol ExerciseEditViewDelegate: AnyObject {
+    func moveToSelectCategoryView()
 }
 
-class ExerciseEditViewController: BaseViewController {
+class ExerciseEditView: UIView {
     
     // MARK: - Componente
-    private lazy var stackView = UIStackView()
+    lazy var stackView = UIStackView()
     
-    private lazy var URLTextFieldBaseView = UIView()
-    private lazy var URLTextField = UITextField()
-    private lazy var URLErrorLabel = UILabel()
+    lazy var URLTextFieldBaseView = UIView()
+    lazy var URLTextField = UITextField()
+    lazy var URLErrorLabel = UILabel()
     
-    private lazy var categorySelecteBaseView = UIView()
-    private lazy var categorySelectedLabel = UILabel()
+    lazy var categorySelecteBaseView = UIView()
+    lazy var categorySelectedLabel = UILabel()
     
-    private lazy var memoTextBaseView = UIView()
-    private lazy var memoTextView = UITextView()
+    lazy var memoTextBaseView = UIView()
+    lazy var memoTextView = UITextView()
     
     // MARK: - 변수
-    private var viewModel = ExerciseEditViewModel()
-    private var categoryViewModel = SelectCategoryViewModel()
-    private var disposeBag = DisposeBag()
-
-    // MARK: - Life Cycle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        displayTopView(false)
+    weak var delegate: ExerciseEditViewDelegate?
+    
+    func configure() {
+        setupUI()
     }
     
     // MARK: - Setup UI
-    override func setupUI() {
-        view.addSubview(stackView)
+    func setupUI() {
+        addSubview(stackView)
         
         setupStackView()
+        setupLayout()
     }
     
     private func setupStackView() {
@@ -58,70 +53,7 @@ class ExerciseEditViewController: BaseViewController {
         setupCategorySelect()
         setupMemoTextView()
     }
-    
-    // MARK: - Setup Layout
-    override func setupLayout() {
-        stackView.snp.makeConstraints { make in
-            make.top.equalTo(view.safeAreaLayoutGuide).offset(Padding.leftRightSpacing.rawValue)
-            make.leading.trailing.equalToSuperview().inset(Padding.leftRightSpacing.rawValue)
-            make.bottom.equalTo(view.safeAreaLayoutGuide.snp.bottom).offset(-Padding.leftRightSpacing.rawValue)
-        }
-    }
-    
-    // MARK: - Setup Delegate
-    override func setupDelegate() {
-    }
-    
-    // MARK: - Setup NavigationBar
-    override func setupNavigationBar() {
-        let button = UIBarButtonItem(title: "저장", style: .plain, target: self, action: #selector(saveData))
-        navigationItem.rightBarButtonItem = button
-        
-        let backButton = UIBarButtonItem(title: "", style: .plain, target: self, action: nil)
-        navigationItem.backBarButtonItem = backButton
-    }
-    
-    @objc func saveData() {
-        let result = viewModel.saveData()
-        let message = result ? "저장했습니다" : "URL 입력 및 카테고리를 선택해 주세요"
-        self.showAlertWithOKButton(title: "", message: message) {
-            if result {
-                self.navigationController?.popViewController(animated: true)
-            }
-        }
-    }
-    
-    // MARK: - Setup Bind
-    override func setupBinding() {
-        URLTextField.rx.text.orEmpty
-            .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
-            .bind(to: viewModel.URLTextField)
-            .disposed(by: disposeBag)
-        
-        viewModel.URLErrorLabel
-            .bind(to: URLErrorLabel.rx.text)
-            .disposed(by: disposeBag)
-        
-        categoryViewModel.selectedCategory
-            .observe(on: MainScheduler.instance)
-            .subscribe { [weak self] category in
-                self?.categorySelectedLabel.text = category?.title ?? "미선택"
-            }
-            .disposed(by: disposeBag)
-        
-        categoryViewModel.selectedCategory
-            .map { $0?.id }
-            .bind(to: viewModel.selectedCategoryId)
-            .disposed(by: disposeBag)
-        
-        memoTextView.rx.text
-            .bind(to: viewModel.memoTextView)
-            .disposed(by: disposeBag)
-    }
-}
 
-// MARK: - Setup 메서드
-extension ExerciseEditViewController {
     private func setupURLTextField() {
         let label = UILabel()
         label.configure(text: "URL" , font: .body)
@@ -188,11 +120,6 @@ extension ExerciseEditViewController {
         }
     }
     
-    @objc func moveToSelectCategoryView() {
-        let viewController = ExerciseSelectCategoryViewController(viewModel: categoryViewModel)
-        navigationController?.pushViewController(viewController, animated: true)
-    }
-    
     private func setupMemoTextView() {
         let label = UILabel()
         label.configure(text: "메모", font: .body)
@@ -215,5 +142,19 @@ extension ExerciseEditViewController {
             make.leading.trailing.equalTo(label)
             make.bottom.equalToSuperview().offset(-24)
         }
+    }
+    
+    // MARK: - Setup Layout
+    func setupLayout() {
+        stackView.snp.makeConstraints { make in
+            make.top.equalTo(safeAreaLayoutGuide).offset(Padding.leftRightSpacing.rawValue)
+            make.leading.trailing.equalToSuperview().inset(Padding.leftRightSpacing.rawValue)
+            make.bottom.equalTo(safeAreaLayoutGuide.snp.bottom).offset(-Padding.leftRightSpacing.rawValue)
+        }
+    }
+    
+    // MARK: - objc 메서드
+    @objc func moveToSelectCategoryView() {
+        delegate?.moveToSelectCategoryView()
     }
 }
